created separate file for each function, with public_api.ts
