+++ src/change-detection-key.ts
@@ -1,0 +1,3 @@
+import * as uuid from 'uuid';
+
+export const changeDetectionKey = uuid();
